Move diary views to own package
